Convert Challenge to Groovy
